Move RssFeed into a separate file
