+++ src/itemlistformaction.cpp
@@ -11,6 +11,7 @@
 #include "fmtstrformatter.h"
 #include "logger.h"
 #include "matcherexception.h"
+#include "rssfeed.h"
 #include "strprintf.h"
 #include "utils.h"
 #include "view.h"
