+++ src/reloader.cpp
@@ -12,6 +12,7 @@
 #include "reloadrangethread.h"
 #include "reloadthread.h"
 #include "rss/exception.h"
+#include "rssfeed.h"
 #include "rssparser.h"
 #include "utils.h"
 #include "view.h"
