+++ src/view.cpp
@@ -40,12 +40,13 @@ extern "C" {
 #include "itemlistformaction.h"
 #include "itemview.h"
 #include "itemviewformaction.h"
-#include "matcherexception.h"
 #include "keymap.h"
 #include "logger.h"
+#include "matcherexception.h"
 #include "regexmanager.h"
 #include "reloadthread.h"
 #include "rss.h"
+#include "rssfeed.h"
 #include "selectformaction.h"
 #include "selecttag.h"
 #include "strprintf.h"
