+++ test/cache.cpp
@@ -4,6 +4,7 @@
 
 #include "3rd-party/catch.hpp"
 #include "configcontainer.h"
+#include "rssfeed.h"
 #include "rssparser.h"
 #include "test-helpers.h"
