+++ test/feedcontainer.cpp
@@ -6,6 +6,7 @@
 #include "configcontainer.h"
 #include "feedcontainer.h"
 #include "rss.h"
+#include "rssfeed.h"
 #include "test-helpers.h"
 
 using namespace newsboat;
