+++ test/itemrenderer.cpp
@@ -3,6 +3,7 @@
 #include "cache.h"
 #include "configcontainer.h"
 #include "itemrenderer.h"
+#include "rssfeed.h"
 #include "test-helpers.h"
 
 using namespace newsboat;
